Implemented better JWT parsing (without any dependency other than bouncy) Implemented connection encryption Implemented various AI behaviours (there is still a bug in the pathfinder) Some random improvements overall

+++ gomint-server/src/main/java/io/gomint/server/entity/metadata/MetadataVector.java
@@ -56,9 +56,9 @@ public class MetadataVector extends MetadataValue {
     @Override
     void serialize( PacketBuffer buffer, int index ) {
         super.serialize( buffer, index );
-        buffer.writeLFloat( (float) this.value.getX() );
-        buffer.writeLFloat( (float) this.value.getY() );
-        buffer.writeLFloat( (float) this.value.getZ() );
+        buffer.writeLFloat( this.value.getX() );
+        buffer.writeLFloat( this.value.getY() );
+        buffer.writeLFloat( this.value.getZ() );
     }
 
     @Override
